improved handling of patchers args
git-svn-id: https://svn.grrrr.org/ext/trunk@796 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/flbase.cpp
@@ -119,7 +119,7 @@ bool flext_obj::GetParamSym(t_atom &dst,const t_symbol *sym,t_canvas *c)
 		while(*c && (isdigit(*c) || *c == '.')) ++c;
 
 		if(!*c) 
-			SetFloat(dst,atof(GetString(res)));
+			SetFloat(dst,(float)atof(GetString(res)));
 		else
 			SetSymbol(dst,res);
 		return true;
